Move SSZ macros into own file
